\name{getColor}
\alias{getColor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getColor
}
\description{
associates a R color name (i.e. in the ouput of colors()) to each possible integer input index. Colors are chosen in a reduced, well differentiated, subset.
}
\usage{
getColor(index)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{index}{
integer input index.
}
}
\value{
color name.
}
\author{
Pierrick Bruneau
}
\examples{
getColor(3)
}

