\name{getTimestamp}
\alias{getTimestamp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getTimestamp
}
\description{
returns current timestamp.
}
\usage{
getTimestamp()
}
\value{
numeric vector with seconds and milliseconds since epoch.
}
\author{
Pierrick Bruneau
}
\seealso{
getElapsed
}
\examples{
temp <- getTimestamp()
}

