\name{handlab}
\alias{handlab}
\docType{data}
\title{
handlab
}
\description{
vector of numeric labels associated to handdat.
}
\format{
  The format is:
 int [1:2000] 0 3 2 0 8 1 3 7 3 7 ...
}
\source{
http://yann.lecun.com/exdb/mnist/
}
\references{
LeCun, Y., Bottou, L., Bengio, Y., and Haffner, P. (1998) _Gradient-based learning applied to document
recognition_, Proceedings of the IEEE, Volume 86, Number 11, Pages 2278-2324.
}
\examples{
handlab[1:10]
}
\keyword{datasets}

