\name{imglabels}
\alias{imglabels}
\docType{data}
\title{
imglabels
}
\description{
vector of numeric labels, indicating the sub-directory in the Caltech-256 collection associated to respective elements in imgmods.
}
\format{
  The format is:
 num [1:1243] 1 1 1 1 1 1 1 1 1 1 ...
}
\examples{
imglabels[1:10]
}
\keyword{datasets}

