\name{irisdata}
\alias{irisdata}
\docType{data}
\title{
irisdata
}
\description{
matrix 150 x 4 of row-elements, extracted from iris standard data.frame (4 first variables). See reference.
}
\format{
  The format is:
 num [1:150, 1:4] 5.1 4.9 4.7 4.6 5 5.4 4.6 5 4.4 4.9 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4] "Sepal.Length" "Sepal.Width" "Petal.Length" "Petal.Width"
}
\references{
Fisher, R. A. (1936) _The use of multiple measurements in taxonomic problems_, Annals of Eugenics, Volume 7, Part II, Pages 179-188.
}
\examples{
displayScatter(irisdata)
}
\keyword{datasets}

