\name{mySmoothScatter}
\alias{mySmoothScatter}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mySmoothScatter
}
\description{
Personalized version of smoothScatter. For internal usage in graphical functions.
}
\usage{
mySmoothScatter(data, model, xlim, ylim)
}
\arguments{
  \item{data}{
matrix of row-elements to plot.
}
  \item{model}{
Optional Gaussian components to plot.
}
  \item{xlim}{
optional bound for plotting.
}
  \item{ylim}{
optional bound for plotting.
}
}

