\name{norm}
\alias{norm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
norm
}
\description{
computes Euclidian norm of vec.
}
\usage{
norm(vec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vec}{
numeric vector.
}
}
\value{
norm value.
}
\author{
Pierrick Bruneau
}
\examples{
temp <- norm(gmmpen[[2]]$mean[[1]])
}

