\name{pcapen}
\alias{pcapen}
\docType{data}
\title{
pcapen
}
\description{
list of 20 MPPCA posterior objects, estimated on subsets of the original 10992-elements pendat data set.
}
\format{
  The format is:
List of 20 posterior MPPCA objects
}
\examples{
temp <- mppcaToGmm(pcapen[[1]])
}
\keyword{datasets}

