\name{pendat}
\alias{pendat}
\docType{data}
\title{
pendat
}
\description{
matrix 2000 x 16 of real row-elements.
}
\format{
  The format is:
 num [1:2000, 1:16] -4.6 -1.2 -2.4 8.4 0.6 3.8 -10 8.8 -10 4.4 ...
}
\source{
http://archive.ics.uci.edu/ml/datasets/Pen-Based+Recognition+of+Handwritten+Digits
}
\references{
Alimoglu, F. (1996) _Combining multiple classifiers for pen-based handwritten digit recognition_, Technical
report, Institute of Graduate Studies in Science and Engineering.
}
\examples{
displayScatter(pendat)
}
\keyword{datasets}

