\name{penlab}
\alias{penlab}
\docType{data}
\title{
penlab
}
\description{
vector of numeric labels associated to pendat.
}
\format{
  The format is:
 int [1:2000] 5 3 8 6 0 9 1 8 1 9 ...
}
\source{
http://archive.ics.uci.edu/ml/datasets/Pen-Based+Recognition+of+Handwritten+Digits
}
\references{
Alimoglu, F. (1996) _Combining multiple classifiers for pen-based handwritten digit recognition_, Technical
report, Institute of Graduate Studies in Science and Engineering.
}
\examples{
displayScatter(data=pendat, labels=penlab)
}
\keyword{datasets}

