\name{Rdct}
\alias{Rdct}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Rdct
}
\description{
performs DCT on a real vector.
}
\usage{
Rdct(vect)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vect}{
vector of real values.
}
}
\value{
vector of DCT transformed values.
}
\author{
Pierrick Bruneau
}
\seealso{
Rdct2D
}
\examples{
temp <- Rdct(irisdata[,1])
}

