\name{vbpen}
\alias{vbpen}
\docType{data}
\title{
vbpen
}
\description{
list of 10 variational posterior GMM objects, estimated on subsets of the original 10992-elements pendat data set.
}
\format{
  The format is:
List of 10 variational GMM.
}
\examples{
temp <- extractSimpleModel(vbpen[[2]])
}
\keyword{datasets}

